Update UpdateLeantime.php
Added a check of the retval of exec('cp') in case, like in 3.3.1, it fails for some reason.  In 3.3.1, it was trying to cp from .../cache/leantime/* but the new files were unzipped into the folder .../cache/*
Also, added a try/catch block around ZipArchive because it was not installed by default on my system.  With the try/catch block, a nice error message can be displayed for the user